requirements.txt
openai>=1.0.0
PyMuPDF>=1.24.0
rich>=13.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
networkx>=3.0
prompt_toolkit>=3.0.0
# Optional: graspologic>=3.0.0 (Leiden algorithm; falls back to networkx Louvain)
